+++ ai/app/src/main/java/com/google/firebase/quickstart/ai/ui/navigation/Navigation.kt
@@ -1,0 +1,13 @@
+package com.google.firebase.quickstart.ai.ui.navigation
+
+import kotlinx.serialization.Serializable
+
+// Within Chat
+@Serializable
+data class TextSample(
+    val initialPrompt: String = "",
+)
+
+// Within Media
+
+// Live
